Move file matches state to separate package
