+++ web/src/collection/state/actions.js
@@ -49,60 +49,6 @@ export function fetchFilesFailure(error) {
 }
 
 /**
- * Single file matches actions
- */
-
-export const ACTION_UPDATE_FILE_MATCH_FILTERS =
-  "coll.UPDATE_FILE_MATCH_FILTERS";
-
-export function updateFileMatchFilters(filters) {
-  return { filters, type: ACTION_UPDATE_FILE_MATCH_FILTERS };
-}
-
-export const ACTION_UPDATE_FILE_MATCH_FILTERS_SUCCESS =
-  "coll.UPDATE_FILE_MATCH_FILTERS_SUCCESS";
-
-export function updateFileMatchFiltersSuccess(matches, files, total) {
-  return {
-    matches,
-    files,
-    total,
-    type: ACTION_UPDATE_FILE_MATCH_FILTERS_SUCCESS,
-  };
-}
-
-export const ACTION_UPDATE_FILE_MATCH_FILTERS_FAILURE =
-  "coll.UPDATE_FILE_MATCH_FILTERS_FAILURE";
-
-export function updateFileMatchFiltersFailure(error) {
-  return { type: ACTION_UPDATE_FILE_MATCH_FILTERS_FAILURE, error };
-}
-
-/**
- * Fetch next matches page
- */
-
-export const ACTION_FETCH_FILE_MATCHES = "coll.FETCH_FILE_MATCHES";
-
-export function fetchFileMatches() {
-  return { type: ACTION_FETCH_FILE_MATCHES };
-}
-
-export const ACTION_FETCH_FILE_MATCHES_SUCCESS =
-  "coll.FETCH_FILE_MATCHES_SUCCESS";
-
-export function fetchFileMatchesSuccess(matches, files, total) {
-  return { matches, files, total, type: ACTION_FETCH_FILE_MATCHES_SUCCESS };
-}
-
-export const ACTION_FETCH_FILE_MATCHES_FAILURE =
-  "coll.FETCH_FILE_MATCHES_FAILURE";
-
-export function fetchFileMatchesFailure(error) {
-  return { error, type: ACTION_FETCH_FILE_MATCHES_FAILURE };
-}
-
-/**
  * File cluster actions
  */
