+++ web/src/collection/state/sagas.js
@@ -1,74 +1,20 @@
-import { call, put, select, takeLatest } from "redux-saga/effects";
+import { fork, call, put, select, takeLatest } from "redux-saga/effects";
 import {
   ACTION_FETCH_FILE_CLUSTER,
-  ACTION_FETCH_FILE_MATCHES,
   ACTION_FETCH_FILES,
   ACTION_UPDATE_FILE_CLUSTER_FILTERS,
-  ACTION_UPDATE_FILE_MATCH_FILTERS,
   ACTION_UPDATE_FILTERS,
   fetchFileClusterFailure,
   fetchFileClusterSuccess,
-  fetchFileMatchesFailure,
-  fetchFileMatchesSuccess,
   fetchFilesFailure,
   fetchFilesSuccess,
   updateFileClusterFiltersFailure,
   updateFileClusterFiltersSuccess,
-  updateFileMatchFiltersFailure,
-  updateFileMatchFiltersSuccess,
   updateFiltersFailure,
   updateFiltersSuccess,
 } from "./actions";
 import { selectColl, selectFileCluster, selectFileMatches } from "./selectors";
-
-function* updateFileMatchesFiltersSaga(server, action) {
-  yield* fetchFileMatchesSaga(
-    server,
-    action,
-    updateFileMatchFiltersSuccess,
-    updateFileMatchFiltersFailure
-  );
-}
-
-function* fetchFileMatchesPageSaga(server, action) {
-  yield* fetchFileMatchesSaga(
-    server,
-    action,
-    fetchFileMatchesSuccess,
-    fetchFileMatchesFailure
-  );
-}
-
-function* fetchFileMatchesSaga(server, action, success, failure) {
-  try {
-    // Determine current query params
-    const { limit, filters, matches: current } = yield select(
-      selectFileMatches
-    );
-
-    // Send request to the server
-    const resp = yield call([server, server.fetchFileMatches], {
-      limit,
-      offset: current.length,
-      id: filters.fileId,
-      filters,
-    });
-
-    // Handle error
-    if (resp.failure) {
-      console.error("Fetch file matches error", resp.error);
-      yield put(failure(resp.error));
-      return;
-    }
-
-    // Update state
-    const { total, matches, files } = resp.data;
-    yield put(success(matches, files, total));
-  } catch (error) {
-    console.error(error);
-    yield put(failure(error));
-  }
-}
+import fileMatchRootSaga from "./fileMatches/sagas";
 
 function* updateFileClusterFiltersSaga(server, action) {
   yield* fetchFileClusterSaga(
@@ -170,7 +116,8 @@ function* fetchFilesSaga(server, action) {
  * Initialize collection-related sagas...
  */
 export default function* collRootSaga(server) {
-  console.log("coll root saga");
+  yield fork(fileMatchRootSaga, server, selectFileMatches);
+
   yield takeLatest(
     [ACTION_UPDATE_FILTERS, ACTION_FETCH_FILES],
     fetchFilesSaga,
@@ -182,11 +129,4 @@ export default function* collRootSaga(server) {
     server
   );
   yield takeLatest(ACTION_FETCH_FILE_CLUSTER, fetchFileClusterPageSaga, server);
-
-  yield takeLatest(
-    ACTION_UPDATE_FILE_MATCH_FILTERS,
-    updateFileMatchesFiltersSaga,
-    server
-  );
-  yield takeLatest(ACTION_FETCH_FILE_MATCHES, fetchFileMatchesPageSaga, server);
 }
